Moved rescore engine into PerFileRescoreTask; dropped AnalysisPipeline inheritance
- Moved RunWorker, ExecuteRescore, WriteReconciledParquet, LoadSpectraForRescore, LoadMassCalibrations, LoadOriginalRtCalibration, AddIfNotNull, and RescoreStats from AnalysisPipeline.PostReconciliationRescore.cs into PerFileRescoreTask; deleted the partial file
- Added a parameterless constructor for the --join-at-pass=1 --no-join worker entry; RescoreWorker.Run now constructs PerFileRescoreTask directly
- Dropped the AnalysisPipeline : AbstractScoringTask inheritance and the Name/Run(ctx) stubs; AnalysisPipeline is now a plain (non-partial) 220-line driver
- Switched LogInfo/LogWarning/Program.LogX call sites in the moved methods to _ctx.LogInfo/_ctx.LogWarning/_ctx.LogError; retargeted the test reference AnalysisPipeline.s_calXcorrScorer to AbstractScoringTask.s_calXcorrScorer
- Snapshot regression PASS at every stage on Stellar AND Astral 3-file; 303/303 OspreySharp unit tests pass

See ai/todos/active/TODO-20260509_osprey_pipeline_tasks.md
