+++ pwiz_tools/OspreySharp/OspreySharp.Test/CalibrationTest.cs
@@ -525,7 +525,7 @@ namespace pwiz.OspreySharp.Test
         /// LOH and trigger gen-2 GC pressure during the all-windows-at-once
         /// pre-preprocessing step.
         ///
-        /// If a future change makes AnalysisPipeline.s_calXcorrScorer
+        /// If a future change makes AbstractScoringTask.s_calXcorrScorer
         /// construct with BinConfig.HRAM() (or any other non-unit bin config),
         /// these asserts fail loudly and cite the Rust design doc. Mirrors
         /// the calibration_scorer_uses_unit_bins_for_* tests in Rust osprey
@@ -534,7 +534,7 @@ namespace pwiz.OspreySharp.Test
         [TestMethod]
         public void TestCalibrationXcorrScorerUsesUnitBins()
         {
-            int calBins = AnalysisPipeline.s_calXcorrScorer.BinConfig.NBins;
+            int calBins = AbstractScoringTask.s_calXcorrScorer.BinConfig.NBins;
             int unitBins = new SpectralScorer(BinConfig.UnitResolution()).BinConfig.NBins;
             int hramBins = new SpectralScorer(BinConfig.HRAM()).BinConfig.NBins;
